Apply interface and impl pattern
