+++ src/nn/OperandImpl.ts
@@ -1,0 +1,3 @@
+import {Operand as OperandInterface} from './Operand';
+
+export class Operand implements OperandInterface {}
