+++ src/nn/types.ts
@@ -1,0 +1,3 @@
+/** */
+export type ArrayBufferView =
+    Float32Array|Int32Array|Uint32Array|Int16Array|Uint16Array;
